Support `nice_partials`
Check for [nice_partials][] support by detecting the `np` helper method.
When present, `yield p = np`, then attempt to read the `:content` and
`:turbo_streams` blocks. If they're present, use them in the partial. If
they're absent, fallback to the `content:` option for `content`, and the
`yield`-ed block content for the `turbo_streams`.

[nice_partials]: https://github.com/bullet-train-co/nice_partials